services/catalog/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-logging==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0

